+++ wxraytracer/wxRaytracer/raytracer/GeometricObjects/Sphere.h
@@ -34,7 +34,7 @@ public:
 
 	virtual bool hit(const Ray& ray, double& tmin, ShadeRec& sr) const;
 
-	virtual bool shadow_hit(const Ray& ray, float& tmin) const;
+	virtual bool shadow_hit(const Ray& ray, double& tmin) const;
 
 private:
